RDKCOM-3176: REFPLTB-1636: add support for mcast2ucast
Reason for change: It is beneficial for debugging process to know iff
multicast-to-unicast conversion is enabled or not and to have
an ability to disable it during debugging to narrow down causes of
potential problems.
This flag can be controlled by the OpenSync framework.
Test Procedure: None
Risks: Low
Signed-off-by: Michal Kowalczyk <mkowalczyk@plume.com>

(cherry picked from commit 06284264c50e1df6730c81e074177979dfcf922f)
Change-Id: I72803a471c3b7d415051ccc91568e23285496bcd